# confero-api/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Confero API
  description: |-
    Some description
  version: 1.0.0
servers:
  - url: api
security:
  - BearerAuth: [ ]
...
paths:
...
  /profile/info:
    put:
      tags:
        - Profile
        - User
      summary: Update user info.
      description: |-
        Update user info.
      operationId: updateUserInfo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileInfoRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/email:
    put:
      tags:
        - Profile
        - User
      summary: Update user email
      description: |-
        Update user email.
      operationId: updateUserEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmailRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/orcid/verify:
    get:
      tags:
        - Profile
        - User
      summary: Verify ORCID
      description: |-
        Verify ORCID.
      operationId: verifyOrcid
      parameters:
        - name: orcid
          in: query
          required: true
          schema:
            type: string
        - name: access_token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/avatar:
    put:
      tags:
        - Profile
        - User
      summary: Update profile avatar
      description: |-
        Update profile avatar.
      operationId: updateAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateAvatarRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/event:
    get:
      tags:
        - Session
        - User
      summary: Get events from all sessions
      operationId: getSessionEvents
      parameters:
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 5

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SesssionEventResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/agenda/add-all-by-organizer/{organizerId}:
    post:
      tags:
        - Session
        - User
      summary: Add all sessions by organizer to user's agenda
      description: Add all sessions by the organizer to the user's personal agenda.
      operationId: addAllSessionsByOrganizerToAgenda
      parameters:
        - $ref: '#/components/parameters/OrginizerId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                description: Number of sessions added to the agenda
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/agenda:
    get:
      tags:
        - Session
        - User
      summary: Get user's conference agenda
      description: |-
        Retrieve all sessions that a user has added to their agenda. Organizers will also see sessions they are involved in, even if they have not yet configured their agenda.
      operationId: getPersonalAgenda
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConferenceAgenda'
    post:
      tags:
        - Session
        - User
      summary: Add a session to user's agenda
      description: Add a single session to the user's personal agenda.
      operationId: addSessionToAgenda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: integer
                  format: int64
                  description: The ID of the session to be added.
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Session
        - User
      summary: Remove a session from user's agenda
      description: Remove a single session from the user's personal agenda.
      operationId: removeSessionFromAgenda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: integer
                  format: int64
                  description: The ID of the session to be deleted from agenda.
      responses:
        '204':
          description: No content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}:
    get:
      tags:
        - Session
        - User
      summary: Get a session
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Session
        - User
      summary: Update a session
      operationId: updateSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/preview:
    get:
      tags:
        - Session
      summary: Get a session for unauthenticated users
      operationId: getSessionPreview
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/presentation/{presentationId}/attachment:
    post:
      tags:
        - Session
        - User
        - Presentation
      summary: Add presentation attachement.
      description: |-
        Add attachement to the presentation. It will be possible only for presentation participants.
      operationId: addPresentationAttachment
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/PresentationId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttachmentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/presentation/{presentationId}/attachment/{attachmentId}:
    delete:
      tags:
        - Session
        - User
        - Presentation
      summary: Delete presentation attachment
      description: |-
        Delete attachment from the presentation. It will be possible only for presentation participants.
      operationId: deletePresentationAttachment
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/AttachmentId'
        - $ref: '#/components/parameters/PresentationId'
      responses:
        '204':
          description: No content
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Session ID
    PresentationId:
      name: presentationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Presentation ID
...
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Attachment ID
...
    OrginizerId:
      name: organizerId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Organizer ID
...
  schemas:
...
    ErrorReason:
      type: string
      enum:
        - ACTIVE_CONFERENCE_EDITION_ALREADY_EXISTS
        - USER_NOT_FOUND
        - APPLICATION_NOT_FOUND
        - SESSION_NOT_FOUND
        - NOT_FOUND
        - INVALID_FILE_FORMAT
        - S3_UPLOAD_ERROR
        - EMAIL_CANNOT_BE_UPDATED
        - NO_ACTIVE_CONFERENCE_EDITION
        - UNAUTHORIZED
        - INVALID_ORCID
        - UNEXPECTED_ERROR
        - ADMIN_CANNOT_CREATE_APPLICATION
        - ADMIN_CANNOT_DELETE_APPLICATION
        - ADMIN_CANNOT_UPDATE_APPLICATION
        - EMAIL_ALREADY_EXISTS
        - EMAIL_SENDING_ERROR
        - NO_PRESENTATIONS_PROVIDED
        - NO_PRESENTERS_PROVIDED
        - FORBIDDEN
        - SESSION_IS_NOT_FROM_CURRENT_CONFERENCE_EDITION
        - PRESENTATION_NOT_FOUND
        - BOTH_START_AND_END_TIME_MUST_BE_PROVIDED_FOR_PRESENTATION_UPDATE
        - ONLY_PARTICIPANTS_CAN_UPDATE_SESSION
        - START_TIME_MUST_BE_BEFORE_END_TIME
        - PRESENTER_NOT_FOUND
        - ONLY_PARTICIPANT_CAN_UPDATE_PRESENTATION
        - ATTACHMENT_NOT_FOUND
        - CONFERENCE_EDITION_CANNOT_HAVE_DEADLINE_IN_THE_PAST
    SessionEventType:
      type: string
      enum: [ FILE_UPLOAD ]
...
    ErrorResponse:
      type: object
      properties:
        reason:
          $ref: '#/components/schemas/ErrorReason'
...
    PresenterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        orcid:
          type: string
        isSpeaker:
          type: boolean
        email:
          type: string
...
    PresentationSessionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentResponse'
        presenters:
          type: array
          items:
            $ref: '#/components/schemas/PresenterResponse'
        isMine:
          type: boolean
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    UpdateSessionRequest:
      type: object
      properties:
        presentations:
          type: array
          items:
            $ref: '#/components/schemas/UpdatePresentationRequest'
    UpdatePresentationRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
...
    SessionPreviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        tags:
          type: array
          items:
            type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        presenters:
          type: array
          items:
            $ref: '#/components/schemas/PresenterResponse'
        isInCalendar:
          type: boolean
        from_current_conference_edition:
          type: boolean
    SessionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        isMine:
          type: boolean
        description:
          type: string
        presentations:
          type: array
          items:
            $ref: '#/components/schemas/PresentationSessionResponse'
        tags:
          type: array
          items:
            type: string
        from_current_conference_edition:
          type: boolean
    AttachmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        url:
          type: string
    AttachmentRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
...
    ConferenceAgenda:
      type: array
      items:
        $ref: '#/components/schemas/SessionPreviewResponse'
    ProfileResponse:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        emails:
          type: array
          items:
            $ref: '#/components/schemas/EmailResponse'
        orcid:
          type: string
        isAdmin:
          type: boolean
        avatarUrl:
          type: string
        isInvitee:
          type: boolean
    EmailResponse:
      type: object
      properties:
        email:
          type: string
        confirmed:
          type: boolean
    UpdateEmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
    UpdateProfileInfoRequest:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
    UpdateAvatarRequest:
      type: object
      properties:
        avatarFile:
          type: string
          format: binary
          description: "Avatar image file to be uploaded"
    SesssionEventResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/SessionEventType'
        userFirstName:
          type: string
        userLastName:
          type: string
        timestamp:
          type: string
          format: date-time
        sessionId:
          type: integer
          format: int64